+++ resources/views/admin/api/new.blade.php
@@ -26,18 +26,30 @@
                 <td scope="col">2</td>
                 <td scope="col">Magento2</td>
                 <td scope="col">
-                    <form class="row row-cols-lg-auto g-3 align-items-center" method="post"
-                        action="{{ route('api.new', ['type' => 'magento2']) }}">
+                    <form method="post" action="{{ route('api.new', ['type' => 'magento2']) }}">
                         @method('POST') @csrf
-                        <div class="col-12">
-                            <label class="visually-hidden" for="apiMagento2Url">Url</label>
-                            <div class="input-group">
-                                <div class="input-group-text">Url</div>
-                                <input type="text" class="form-control" id="apiMagento2Url"
-                                    placeholder="https://www.domain.com">
+                        <x-row class="row row-cols-lg-auto g-3 align-items-center mb-1">
+                            <div class="col-12">
+                                <input class="form-control" name="url" placeholder="https://www.domain.com">
                             </div>
-                        </div>
-                        <div class="col-12">
+                        </x-row>
+                        <x-row class="row-cols-lg-auto g-3 align-items-center mb-1">
+                            <div class="col-12">
+                                <input class="form-control" name="oauth_consumer_key" placeholder="Consumer Key">
+                            </div>
+                            <div class="col-12">
+                                <input class="form-control" name="oauth_consumer_secret" placeholder="Consumer Secret">
+                            </div>
+                        </x-row>
+                        <x-row class="row-cols-lg-auto g-3 align-items-center mb-1">
+                            <div class="col-12">
+                                <input class="form-control" name="oauth_access_token" placeholder="Access Token">
+                            </div>
+                            <div class="col-12">
+                                <input class="form-control" name="oauth_access_secret"
+                                    placeholder="Access Token Secret">
+                            </div>
+                            {{-- <div class="col-12">
                             <label class="visually-hidden" for="apiMagento2Username">Username</label>
                             <div class="input-group">
                                 <div class="input-group-text">Acc</div>
@@ -52,10 +64,13 @@
                                 <input type="text" class="form-control" id="apiMagento2Password"
                                     placeholder="Password">
                             </div>
-                        </div>
-                        <div class="col-12">
-                            <button type="submit" class="btn btn-success">Get Api</button>
-                        </div>
+                        </div> --}}
+                        </x-row>
+                        <x-row class="mb-1">
+                            <div class="col-12">
+                                <button type="submit" class="btn btn-success">Get Api</button>
+                            </div>
+                        </x-row>
                     </form>
                 </td>
             </tr>