+++ resources/views/catalog/products.blade.php
@@ -25,7 +25,7 @@
                 <x-row class="row-cols-auto">
                     @foreach ($apis as $api)
                         <div class="col-auto">
-                            <button type="submit" name="api[]" value="{{ $api->id }}"
+                            <button type="submit" name="api" value="{{ $api->id }}"
                                 class="btn btn-primary mb-3">
                                 Đồng bộ từ {{ $api->type }}
                             </button>